+++ wpull/converter.py
@@ -1,11 +1,16 @@
 # encoding=utf-8
+'''Document content post-processing.'''
 import abc
 
 
 class BaseDocumentConverter(object, metaclass=abc.ABCMeta):
-    pass
+    '''Base class for classes that convert a document format.'''
+    @abc.abstractmethod
+    def convert(self, input_file, output_file):
+        pass
 
 
 class LocalHTMLConverter(BaseDocumentConverter):
     # TODO: convert links to local
-    pass
+    def convert(self, input_file, output_file):
+        raise NotImplementedError()